Fast forward to latest trunk revision 3788.

+++ branches/symbolic_from_3470/pycad/py_src/Triangle.py
@@ -54,6 +54,7 @@ class Design(design.Design):
        if dim != 2:
            raise ValueError("only dimension 2 is supported by Triangle.")
        design.Design.__init__(self,dim=dim,keep_files=keep_files)
+       self.__scriptname=""
        self.setScriptFileName()
        self.setMeshFileName()
        self.setOptions()
@@ -63,12 +64,15 @@ class Design(design.Design):
        Sets the filename for the Triangle input script. If no name is given
        a name with extension I{poly} is generated.
        """
+       if self.__scriptname:
+           os.unlink(self.__scriptname)
        if name == None:
+           self.__scriptname_set=False
            self.__scriptname=tempfile.mkstemp(suffix=".poly")[1]
        else:
+           self.__scriptname_set=True
            self.__scriptname=name
            self.setMeshFileName(name)
-           self.setKeepFilesOn()
 
     def getScriptFileName(self):
        """
@@ -80,6 +84,8 @@ class Design(design.Design):
        """
        Sets the name of the Triangle mesh file.
        """
+       if self.__mshname:
+           os.unlink(self.__mshname)
        if name == None:
            self.__mshname=tempfile.mkstemp(suffix="")[1]
        else:
@@ -116,8 +122,10 @@ class Design(design.Design):
         Cleans up.
         """
         if not self.keepFiles():
-               os.unlink(self.getScriptFileName())
-               os.unlink(self.getMeshFileName())
+            if not self.__scriptname_set:
+                os.unlink(self.getScriptFileName())
+            if not self.__mshname_set:
+                os.unlink(self.getMeshFileName())
 
     def getCommandString(self):
         """
@@ -128,7 +136,7 @@ class Design(design.Design):
             see U{http://www.cs.cmu.edu/~quake/triangle.switch.html}
         """
         if self.__cmdLineArgs == "":
-            print "warning: using default command line arguments for Triangle"
+            print("warning: using default command line arguments for Triangle")
         exe="triangle %s %%s"%self.__cmdLineArgs
         return exe
 
@@ -147,7 +155,7 @@ class Design(design.Design):
             ret=0
         ret=getMPIWorldMax(ret)
         if ret > 0:
-          raise RuntimeError, "Could not build mesh: %s"%" ".join(args)
+          raise RuntimeError("Could not build mesh: %s"%" ".join(args))
         else:
             # <hack> so that users can set the mesh filename they want.
             name=self.getScriptFileName()
@@ -192,7 +200,7 @@ class Design(design.Design):
                        pts=list(PSp.getControlPoints())
                        for pt in pts:
                            c=pt.getCoordinates()
-                           if pt not in ctrlPts.keys():
+                           if pt not in list(ctrlPts.keys()):
                                vertCnt+=1
                                vertices+="%s %s %s %s\n"%(vertCnt,c[0],c[1],p.getID())
                                ctrlPts[pt]=vertCnt
@@ -217,7 +225,7 @@ class Design(design.Design):
                            pts=list(curve.getControlPoints())
                            for pt in pts:
                                c=pt.getCoordinates()
-                               if pt not in ctrlPts.keys():
+                               if pt not in list(ctrlPts.keys()):
                                    vertCnt+=1
                                    vertices+="%s %s %s %s\n"%(vertCnt,c[0],c[1],p.getID())
                                    ctrlPts[pt]=vertCnt
@@ -241,7 +249,7 @@ class Design(design.Design):
                                pts=list(curve.getControlPoints())
                                for pt in pts:
                                    c=pt.getCoordinates()
-                                   if pt not in ctrlPts.keys():
+                                   if pt not in list(ctrlPts.keys()):
                                        vertCnt+=1
                                        vertices+="%s %s %s %s\n"%(vertCnt,c[0],c[1],p.getID())
                                        ctrlPts[pt]=vertCnt
@@ -271,13 +279,13 @@ class Design(design.Design):
                                vectors[i].append(self.__getVector(A,B))
                                vectors[i].append(self.__getVector(A,C))
                            # get angle between vectors at each vertex
-                           for i in vectors.keys():
+                           for i in list(vectors.keys()):
                                angle=self.__getAngle(vectors[i][0],vectors[i][1])
                                vectors[i].append(angle)
                            # find the vertex with the smallest angle
                            minAngle=360.
                            indx=0
-                           for i in vectors.keys():
+                           for i in list(vectors.keys()):
                                if vectors[i][2] < minAngle:
                                    indx=i
                                    minAngle=vectors[i][2]
